requirements.txt
google-auth==2.13.0
google-api-python-client==2.65.0
